+++ cli/build.rs
@@ -309,9 +309,11 @@ mod ts {
 // deps = [runtime]
 deno_core::extension!(
   cli,
+  esm_entry_point = "ext:cli/99_main.js",
   esm = [
     dir "js",
-    "40_testing.js"
+    "40_testing.js",
+    "99_main.js"
   ],
   customizer = |ext: &mut deno_core::ExtensionBuilder| {
     ext.esm(vec![ExtensionFileSource {
