+++ cli/js/99_main.js
@@ -1,0 +1,3 @@
+// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+import "ext:cli/40_testing.js";
+import "ext:cli/runtime/js/99_main.js";
