+++ ext/crypto/lib.rs
@@ -103,7 +103,7 @@ deno_core::extension!(deno_crypto,
     x25519::op_crypto_export_spki_x25519,
     x25519::op_crypto_export_pkcs8_x25519,
   ],
-  esm = [ "00_crypto.js", "01_webidl.js" ],
+  esm = [ "00_crypto.js" ],
   options = {
     maybe_seed: Option<u64>,
   },
